refactor: depreciate Layout.tsx
