+++ src/popup/App.tsx
@@ -1,17 +1,28 @@
 import React, { useEffect, useState } from 'react';
+import Box from '@mui/material/Box';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import { ConfirmProvider } from 'material-ui-confirm';
+import { SnackbarProvider } from 'notistack';
 
 import useTimer from '@hooks/useTimer';
 import useApp from '@stores/useApp';
-import PageLayout from './Layout';
 import useInitializeStore from '../stores/useInitializeStore';
+
+const Player = React.lazy(() => import('../components/App/App'));
 
 export default function App() {
   // The current playing list
   const [currentSongList, setCurrentSongList] = useState<NoxMedia.Song[]>([]);
   const { initializeStores } = useInitializeStore();
+  const playerStyle = useApp((state) => state.playerStyle);
+  const theme = createTheme(playerStyle.colorTheme.palette);
+  const [backgroundSrc, setBackgroundSrc] = React.useState<string>();
   // eslint-disable-next-line no-unused-vars
   const timer = useTimer();
-  const playerStyle = useApp((state) => state.playerStyle);
+
+  if (!currentSongList) {
+    return <h1>Loading...</h1>;
+  }
 
   useEffect(() => {
     async function init() {
@@ -19,18 +30,78 @@ export default function App() {
       setCurrentSongList(result.currentPlayingList.songList);
     }
     init();
-  }, []);
-
-  useEffect(() => {
     document.title = playerStyle.appTitle;
   }, []);
 
+  useEffect(() => {
+    playerStyle.playerBackground().then(setBackgroundSrc);
+  }, [playerStyle.playerBackground]);
+
   // console.log(currentSongList)
   return (
-    <PageLayout
-      songList={currentSongList}
-      // Mobile interface is assumed broken after refactoring and deprecaited
-      // in favor of APM. use the better thing!
-    />
+    // Outmost layer of the page
+    <React.Suspense fallback={<h1>Loading...</h1>}>
+      <ThemeProvider theme={theme}>
+        <SnackbarProvider maxSnack={1}>
+          <ConfirmProvider>
+            <div className='container-fluid homepage-bgimage'>
+              {playerStyle.playerBackgroundVideo ? (
+                <video
+                  id='player-bkgrd'
+                  autoPlay
+                  loop
+                  muted
+                  className='homepage-bgimage'
+                  src={backgroundSrc}
+                  height={window.innerHeight}
+                  width={window.innerWidth}
+                />
+              ) : (
+                <img
+                  id='player-bkgrd'
+                  alt=''
+                  className='homepage-bgimage'
+                  src={backgroundSrc}
+                  height={window.innerHeight}
+                  width={window.innerWidth}
+                />
+              )}
+            </div>
+            <Box
+              sx={OutmostBox}
+              id='master-box'
+              style={{
+                backgroundColor: playerStyle.colorTheme.PCBackgroundColor,
+                backgroundBlendMode: 'overlay',
+              }}
+            >
+              <Box sx={PlayerBox}>
+                <Player songList={currentSongList} />
+              </Box>
+            </Box>
+          </ConfirmProvider>
+        </SnackbarProvider>
+      </ThemeProvider>
+    </React.Suspense>
   );
 }
+
+const OutmostBox = {
+  width: '100vw',
+  height: '95vh',
+  color: '#1234',
+  '& > .MuiBox-root > .MuiBox-root': {
+    p: 1,
+  },
+};
+const PlayerBox = {
+  height: '100vh',
+  maxHeight: '100%',
+  display: 'grid',
+  gridTemplateColumns: 'repeat(4, 1fr)',
+  gap: 0,
+  gridTemplateRows: '72px 1fr',
+  gridTemplateAreas: `"Lrc         Lrc      Lrc      search"
+                        "Lrc         Lrc      Lrc      sidebar"
+                        "footer      footer   footer   footer"`,
+};
